requirements.txt
streamlit
pandas
xhtml2pdf
jinja2
python-dotenv
Pillow
requests
